requirements.txt
MetaTrader5
pandas
python-telegram-bot
python-dotenv
pytz
requests
